+++ include/GafferUI/Gadget.h
@@ -71,6 +71,7 @@ namespace GafferUI
 
 IE_CORE_FORWARDDECLARE( Gadget );
 IE_CORE_FORWARDDECLARE( Style );
+IE_CORE_FORWARDDECLARE( ViewportGadget );
 
 /// Gadgets are zoomable UI elements. They draw themselves using OpenGL, and provide an interface for
 /// handling events. To present a Gadget in the user interface, it should be placed in the viewport of
@@ -87,17 +88,14 @@ class GAFFERUI_API Gadget : public Gaffer::GraphComponent
 
 		enum class Layer
 		{
+			None = -100,
+
 			Back = -2,
 			MidBack = -1,
 			Main = 0,
 			MidFront = 1,
 			Front = 2,
 		};
-
-		/// Returns the Gadget with the specified name, where name has been retrieved
-		/// from an IECoreGL::HitRecord after rendering some Gadget in GL_SELECT mode.
-		/// \todo Consider better mechanisms.
-		static GadgetPtr select( GLuint id );
 
 		/// @name Parent-child relationships
 		////////////////////////////////////////////////////////////////////
@@ -179,13 +177,6 @@ class GAFFERUI_API Gadget : public Gaffer::GraphComponent
 		/// @name Display
 		////////////////////////////////////////////////////////////////////
 		//@{
-		/// Renders the Gadget into the current OpenGL context. If currentStyle
-		/// is passed then it must already have been bound with Style::bind(),
-		/// and will be used if and only if not overridden by a Style applied
-		/// specifically to this Gadget. Typically users will not pass currentStyle -
-		/// but it must be passed by Gadget implementations when rendering child
-		/// Gadgets in doRenderLayer().
-		void render() const;
 		/// The bounding box of the Gadget before transformation. The default
 		/// implementation returns the union of the transformed bounding boxes
 		/// of all the children.
@@ -194,8 +185,6 @@ class GAFFERUI_API Gadget : public Gaffer::GraphComponent
 		Imath::Box3f transformedBound() const;
 		/// The bounding box transformed by the result of fullTransform( ancestor ).
 		Imath::Box3f transformedBound( const Gadget *ancestor ) const;
-		typedef boost::signal<void ( Gadget * )> RenderRequestSignal;
-		RenderRequestSignal &renderRequestSignal();
 		//@}
 
 		/// @name Tool tips
@@ -306,16 +295,15 @@ class GAFFERUI_API Gadget : public Gaffer::GraphComponent
 		/// The default implementation returns true.
 		virtual bool hasLayer( Layer layer ) const;
 
-		/// \deprecated
-		void requestRender();
 		/// Implemented to dirty the layout for both the old and the new parent.
 		void parentChanged( GraphComponent *oldParent ) override;
 
 	private :
+		/// Returns the Gadget with the specified name, where name has been retrieved
+		/// from an IECoreGL::HitRecord after rendering some Gadget in GL_SELECT mode.
+		/// \todo Consider better mechanisms.
+		static GadgetPtr select( GLuint id );
 
-		// Sets the GL state up with the name attribute and transform for
-		// this Gadget, makes sure the style is bound and then calls doRenderLayer().
-		void renderLayer( Layer layer, const Style *currentStyle = nullptr ) const;
 
 		void styleChanged();
 		void emitDescendantVisibilityChanged();
@@ -343,6 +331,9 @@ class GAFFERUI_API Gadget : public Gaffer::GraphComponent
 		// when absolutely necessary (when slots are connected).
 		static IdleSignal &idleSignalAccessedSignal();
 		friend void GafferUIModule::bindGadget();
+
+		/// ViewportGadget performs the actual rendering, and needs access to the internals of all the gadgets it renders
+		friend ViewportGadget;
 
 };