+++ src/GafferUI/GraphGadget.cpp
@@ -682,10 +682,8 @@ NodeGadget *GraphGadget::nodeGadgetAt( const IECore::LineSegment3f &lineInGadget
 {
 	const ViewportGadget *viewportGadget = ancestor<ViewportGadget>();
 
-	std::vector<GadgetPtr> gadgetsUnderMouse;
-	viewportGadget->gadgetsAt(
-		viewportGadget->gadgetToRasterSpace( lineInGadgetSpace.p0, this ),
-		gadgetsUnderMouse
+	std::vector<Gadget*> gadgetsUnderMouse = viewportGadget->gadgetsAt(
+		viewportGadget->gadgetToRasterSpace( lineInGadgetSpace.p0, this )
 	);
 
 	if( !gadgetsUnderMouse.size() )
@@ -693,7 +691,7 @@ NodeGadget *GraphGadget::nodeGadgetAt( const IECore::LineSegment3f &lineInGadget
 		return nullptr;
 	}
 
-	NodeGadget *nodeGadget = runTimeCast<NodeGadget>( gadgetsUnderMouse[0].get() );
+	NodeGadget *nodeGadget = runTimeCast<NodeGadget>( gadgetsUnderMouse[0] );
 	if( !nodeGadget )
 	{
 		nodeGadget = gadgetsUnderMouse[0]->ancestor<NodeGadget>();
@@ -706,15 +704,16 @@ ConnectionGadget *GraphGadget::connectionGadgetAt( const IECore::LineSegment3f &
 {
 	const ViewportGadget *viewportGadget = ancestor<ViewportGadget>();
 
-	std::vector<GadgetPtr> gadgetsUnderMouse;
-	viewportGadget->gadgetsAt( viewportGadget->gadgetToRasterSpace( lineInGadgetSpace.p0, this ), gadgetsUnderMouse );
+	std::vector<Gadget*> gadgetsUnderMouse = viewportGadget->gadgetsAt(
+		viewportGadget->gadgetToRasterSpace( lineInGadgetSpace.p0, this )
+	);
 
 	if ( !gadgetsUnderMouse.size() )
 	{
 		return nullptr;
 	}
 
-	ConnectionGadget *connectionGadget = runTimeCast<ConnectionGadget>( gadgetsUnderMouse[0].get() );
+	ConnectionGadget *connectionGadget = runTimeCast<ConnectionGadget>( gadgetsUnderMouse[0] );
 	if ( !connectionGadget )
 	{
 		connectionGadget = gadgetsUnderMouse[0]->ancestor<ConnectionGadget>();
@@ -725,35 +724,28 @@ ConnectionGadget *GraphGadget::connectionGadgetAt( const IECore::LineSegment3f &
 
 ConnectionGadget *GraphGadget::reconnectionGadgetAt( const NodeGadget *gadget, const IECore::LineSegment3f &lineInGadgetSpace ) const
 {
-	std::vector<GadgetPtr> gadgetsUnderMouse;
-
+	const ViewportGadget *viewportGadget = ancestor<ViewportGadget>();
 	Imath::V3f center = gadget->transformedBound( this ).center();
-	const Imath::V3f corner0 = center - Imath::V3f( 2, 2, 1 );
-	const Imath::V3f corner1 = center + Imath::V3f( 2, 2, 1 );
 
-	std::vector<IECoreGL::HitRecord> selection;
+	Box2f rasterRegion;
+	rasterRegion.extendBy( viewportGadget->gadgetToRasterSpace( center - Imath::V3f( 2, 2, 1 ), this ) );
+	rasterRegion.extendBy( viewportGadget->gadgetToRasterSpace( center + Imath::V3f( 2, 2, 1 ), this ) );
+
+	std::vector<Gadget*> gadgetsUnderMouse = viewportGadget->gadgetsAt(
+		rasterRegion, GraphLayer::Connections
+	);
+	for( Gadget* g : gadgetsUnderMouse )
 	{
-		ViewportGadget::SelectionScope selectionScope( corner0, corner1, this, selection, IECoreGL::Selector::IDRender );
-
-		for ( ChildContainer::const_iterator it = children().begin(); it != children().end(); ++it )
+		if( ConnectionGadget *c = IECore::runTimeCast<ConnectionGadget>( g ) )
 		{
-			if ( ConnectionGadget *c = IECore::runTimeCast<ConnectionGadget>( it->get() ) )
+			if(
+				c->srcNodule() &&
+				gadget->node() != c->srcNodule()->plug()->node() &&
+				gadget->node() != c->dstNodule()->plug()->node()
+			)
 			{
-				// don't consider the node's own connections, or connections without a source nodule
-				if ( c->srcNodule() && gadget->node() != c->srcNodule()->plug()->node() && gadget->node() != c->dstNodule()->plug()->node() )
-				{
-					c->render();
-				}
+				return c;
 			}
-		}
-	}
-
-	for ( std::vector<IECoreGL::HitRecord>::const_iterator it = selection.begin(); it != selection.end(); ++it )
-	{
-		GadgetPtr gadget = Gadget::select( it->name );
-		if ( gadget )
-		{
-			return runTimeCast<ConnectionGadget>( gadget.get() );
 		}
 	}
 
@@ -1025,10 +1017,8 @@ bool GraphGadget::buttonPress( GadgetPtr gadget, const ButtonEvent &event )
 
 		ViewportGadget *viewportGadget = ancestor<ViewportGadget>();
 
-		std::vector<GadgetPtr> gadgetsUnderMouse;
-		viewportGadget->gadgetsAt(
-			viewportGadget->gadgetToRasterSpace( event.line.p0, this ),
-			gadgetsUnderMouse
+		std::vector<Gadget*> gadgetsUnderMouse = viewportGadget->gadgetsAt(
+			viewportGadget->gadgetToRasterSpace( event.line.p0, this )
 		);
 
 		if( !gadgetsUnderMouse.size() || gadgetsUnderMouse[0] == this )
@@ -1042,7 +1032,7 @@ bool GraphGadget::buttonPress( GadgetPtr gadget, const ButtonEvent &event )
 			return true;
 		}
 
-		NodeGadget *nodeGadget = runTimeCast<NodeGadget>( gadgetsUnderMouse[0].get() );
+		NodeGadget *nodeGadget = runTimeCast<NodeGadget>( gadgetsUnderMouse[0] );
 		if( !nodeGadget )
 		{
 			nodeGadget = gadgetsUnderMouse[0]->ancestor<NodeGadget>();