+++ src/GafferUI/ViewportGadget.cpp
@@ -689,11 +689,10 @@ std::string ViewportGadget::getToolTip( const IECore::LineSegment3f &line ) cons
 		return result;
 	}
 
-	std::vector<GadgetPtr> gadgets;
-	gadgetsAt( V2f( line.p0.x, line.p0.y ), gadgets );
-	for( std::vector<GadgetPtr>::const_iterator it = gadgets.begin(), eIt = gadgets.end(); it != eIt; it++ )
+	std::vector<Gadget*> gadgets = gadgetsAt( V2f( line.p0.x, line.p0.y ) );
+	for( Gadget *it : gadgets )
 	{
-		Gadget *gadget = it->get();
+		Gadget *gadget = it;
 		while( gadget && gadget != this )
 		{
 			IECore::LineSegment3f lineInGadgetSpace = rasterToGadgetSpace( V2f( line.p0.x, line.p0.y), gadget );
@@ -921,20 +920,26 @@ bool ViewportGadget::getVariableAspectZoom() const
 	return m_variableAspectZoom;
 }
 
-void ViewportGadget::gadgetsAt( const Imath::V2f &rasterPosition, std::vector<GadgetPtr> &gadgets ) const
+std::vector< Gadget* > ViewportGadget::gadgetsAt( const Imath::V2f &rasterPosition ) const
+{
+	return gadgetsAt( Box2f( rasterPosition - V2f( 1 ), rasterPosition + V2f( 1 ) ) );
+}
+
+std::vector< Gadget* > ViewportGadget::gadgetsAt( const Imath::Box2f &rasterRegion, Gadget::Layer filterLayer ) const
 {
 	std::vector<HitRecord> selection;
 	{
-		SelectionScope selectionScope( this, rasterPosition, selection, IECoreGL::Selector::IDRender );
-		Gadget::render();
+		SelectionScope selectionScope( this, rasterRegion, selection, IECoreGL::Selector::IDRender );
+		renderInternal( filterLayer );
 	}
 
+	std::vector< Gadget* > gadgets;
 	for( std::vector<HitRecord>::const_iterator it = selection.begin(); it!= selection.end(); it++ )
 	{
 		GadgetPtr gadget = Gadget::select( it->name );
 		if( gadget )
 		{
-			gadgets.push_back( gadget );
+			gadgets.push_back( gadget.get() );
 		}
 	}
 
@@ -944,6 +949,17 @@ void ViewportGadget::gadgetsAt( const Imath::V2f &rasterPosition, std::vector<Ga
 		{
 			gadgets.push_back( const_cast<Gadget *>( g ) );
 		}
+	}
+	return gadgets;
+}
+
+// DEPRECATED
+void ViewportGadget::gadgetsAt( const Imath::V2f &rasterPosition, std::vector<GadgetPtr> &gadgets ) const
+{
+	std::vector< Gadget* > retGadgets = gadgetsAt( rasterPosition );
+	for( Gadget *g : retGadgets )
+	{
+		gadgets.push_back( g );
 	}
 }
 
@@ -1007,12 +1023,84 @@ void ViewportGadget::render() const
 	glMultMatrixf( camera->getTransform().getValue() );
 	glMatrixMode( GL_MODELVIEW );
 
-	Gadget::render();
+	renderInternal();
 }
+
+void ViewportGadget::renderInternal( Gadget::Layer filterLayer ) const
+{
+
+	bound(); // Updates layout if necessary
+	for( int layer = (int)Layer::Back; layer <= (int)Layer::Front; ++layer )
+	{
+		if( filterLayer != Gadget::Layer::None && (int)filterLayer != layer )
+		{
+			continue;
+		}
+
+		renderLayer( this, (Layer)layer, /* currentStyle = */ nullptr );
+	}
+}
+
+void ViewportGadget::renderLayer( const Gadget *gadget, Layer layer, const Style *currentStyle )
+{
+	const bool haveTransform = gadget->m_transform != M44f();
+	if( haveTransform )
+	{
+		glPushMatrix();
+		glMultMatrixf( gadget->m_transform.getValue() );
+	}
+
+		if( !currentStyle )
+		{
+			currentStyle = gadget->style();
+			currentStyle->bind();
+		}
+		else
+		{
+			if( gadget->m_style )
+			{
+				gadget->m_style->bind();
+				currentStyle = gadget->m_style.get();
+			}
+		}
+
+		if( IECoreGL::Selector *selector = IECoreGL::Selector::currentSelector() )
+		{
+			selector->loadName( gadget->m_glName );
+		}
+
+		gadget->doRenderLayer( layer, currentStyle );
+
+		for( ChildContainer::const_iterator it=gadget->children().begin(); it!=gadget->children().end(); it++ )
+		{
+			// Cast is safe because of the guarantees acceptsChild() gives us
+			const Gadget *c = static_cast<const Gadget *>( it->get() );
+			if( !c->getVisible() )
+			{
+				continue;
+			}
+			if( c->hasLayer( layer ) )
+			{
+				renderLayer( c, layer, currentStyle );
+			}
+		}
+
+	if( haveTransform )
+	{
+		glPopMatrix();
+	}
+}
+
+
 
 ViewportGadget::UnarySignal &ViewportGadget::preRenderSignal()
 {
 	return m_preRenderSignal;
+}
+
+ViewportGadget::RenderRequestSignal &ViewportGadget::renderRequestSignal()
+{
+	return m_renderRequestSignal;
 }
 
 void ViewportGadget::childRemoved( GraphComponent *parent, GraphComponent *child )
@@ -1041,10 +1129,9 @@ bool ViewportGadget::buttonPress( GadgetPtr gadget, const ButtonEvent &event )
 		return true;
 	}
 
-	std::vector<GadgetPtr> gadgets;
-	gadgetsAt( V2f( event.line.p0.x, event.line.p0.y ), gadgets );
+	std::vector<Gadget*> gadgets = gadgetsAt( V2f( event.line.p0.x, event.line.p0.y ) );
 
-	GadgetPtr handler;
+	Gadget* handler;
 	m_lastButtonPressGadget = nullptr;
 	bool result = dispatchEvent( gadgets, &Gadget::buttonPressSignal, event, handler );
 	if( result )
@@ -1067,7 +1154,7 @@ bool ViewportGadget::buttonRelease( GadgetPtr gadget, const ButtonEvent &event )
 	bool result = false;
 	if( m_lastButtonPressGadget )
 	{
-		result = dispatchEvent( m_lastButtonPressGadget, &Gadget::buttonReleaseSignal, event );
+		result = dispatchEvent( m_lastButtonPressGadget.get(), &Gadget::buttonReleaseSignal, event );
 	}
 
 	m_lastButtonPressGadget = nullptr;
@@ -1083,10 +1170,9 @@ bool ViewportGadget::buttonDoubleClick( GadgetPtr gadget, const ButtonEvent &eve
 
 void ViewportGadget::updateGadgetUnderMouse( const ButtonEvent &event )
 {
-	std::vector<GadgetPtr> gadgets;
-	gadgetsAt( V2f( event.line.p0.x, event.line.p0.y ), gadgets );
+	std::vector<Gadget*> gadgets = gadgetsAt( V2f( event.line.p0.x, event.line.p0.y ) );
 
-	GadgetPtr newGadgetUnderMouse;
+	Gadget* newGadgetUnderMouse = nullptr;
 	if( gadgets.size() )
 	{
 		newGadgetUnderMouse = gadgets[0];
@@ -1144,9 +1230,9 @@ void ViewportGadget::emitEnterLeaveEvents( GadgetPtr newGadgetUnderMouse, Gadget
 			enterTargets.push_back( enterTarget );
 			enterTarget = enterTarget->parent<Gadget>();
 		}
-		for( std::vector<GadgetPtr>::const_reverse_iterator it = enterTargets.rbegin(); it!=enterTargets.rend(); it++ )
+		for( GadgetPtr &it : enterTargets )
 		{
-			dispatchEvent( *it, &Gadget::enterSignal, event );
+			dispatchEvent( it.get(), &Gadget::enterSignal, event );
 		}
 	}
 
@@ -1174,8 +1260,8 @@ bool ViewportGadget::mouseMove( GadgetPtr gadget, const ButtonEvent &event )
 	// pass the signal through
 	if( m_gadgetUnderMouse )
 	{
-		std::vector<GadgetPtr> gadgetUnderMouse( 1, m_gadgetUnderMouse );
-		GadgetPtr handler;
+		std::vector<Gadget*> gadgetUnderMouse( 1, m_gadgetUnderMouse.get() );
+		Gadget* handler;
 		return dispatchEvent( gadgetUnderMouse, &Gadget::mouseMoveSignal, event, handler );
 	}
 
@@ -1192,7 +1278,7 @@ IECore::RunTimeTypedPtr ViewportGadget::dragBegin( GadgetPtr gadget, const DragD
 	if( ( !cameraMotionType || unmodifiedMiddleDrag ) && m_lastButtonPressGadget )
 	{
 		// see if a child gadget would like to start a drag because the camera doesn't handle the event
-		RunTimeTypedPtr data = dispatchEvent( m_lastButtonPressGadget, &Gadget::dragBeginSignal, event );
+		RunTimeTypedPtr data = dispatchEvent( m_lastButtonPressGadget.get(), &Gadget::dragBeginSignal, event );
 		if( data )
 		{
 			const_cast<DragDropEvent &>( event ).sourceGadget = m_lastButtonPressGadget;
@@ -1243,10 +1329,9 @@ bool ViewportGadget::dragEnter( GadgetPtr gadget, const DragDropEvent &event )
 	}
 	else
 	{
-		std::vector<GadgetPtr> gadgets;
-		gadgetsAt( V2f( event.line.p0.x, event.line.p0.y ), gadgets );
+		std::vector<Gadget*> gadgets = gadgetsAt( V2f( event.line.p0.x, event.line.p0.y ) );
 
-		GadgetPtr dragDestination = updatedDragDestination( gadgets, event );
+		Gadget* dragDestination = updatedDragDestination( gadgets, event );
 		if( dragDestination )
 		{
 			const_cast<DragDropEvent &>( event ).destinationGadget = dragDestination;
@@ -1279,14 +1364,13 @@ bool ViewportGadget::dragMove( GadgetPtr gadget, const DragDropEvent &event )
 		// step as an optimisation.
 		if( !event.destinationGadget || !event.data->isInstanceOf( IECore::NullObjectTypeId ) )
 		{
-			std::vector<GadgetPtr> gadgets;
-			gadgetsAt( V2f( event.line.p0.x, event.line.p0.y ), gadgets );
+			std::vector<Gadget*> gadgets = gadgetsAt( V2f( event.line.p0.x, event.line.p0.y ) );
 
 			// update drag destination
-			GadgetPtr updatedDestination = updatedDragDestination( gadgets, event );
+			Gadget* updatedDestination = updatedDragDestination( gadgets, event );
 			if( updatedDestination != event.destinationGadget )
 			{
-				GadgetPtr previousDestination = event.destinationGadget;
+				Gadget* previousDestination = event.destinationGadget.get();
 				const_cast<DragDropEvent &>( event ).destinationGadget = updatedDestination;
 				if( previousDestination )
 				{
@@ -1298,7 +1382,7 @@ bool ViewportGadget::dragMove( GadgetPtr gadget, const DragDropEvent &event )
 		// dispatch drag move to current destination
 		if( event.destinationGadget )
 		{
-			return dispatchEvent( event.destinationGadget, &Gadget::dragMoveSignal, event );
+			return dispatchEvent( event.destinationGadget.get(), &Gadget::dragMoveSignal, event );
 		}
 	}
 
@@ -1450,11 +1534,11 @@ V2f ViewportGadget::motionPositionFromEvent( const DragDropEvent &event ) const
 	}
 }
 
-GadgetPtr ViewportGadget::updatedDragDestination( std::vector<GadgetPtr> &gadgets, const DragDropEvent &event )
+Gadget* ViewportGadget::updatedDragDestination( std::vector<Gadget*> &gadgets, const DragDropEvent &event )
 {
-	for( std::vector<GadgetPtr>::const_iterator it = gadgets.begin(), eIt = gadgets.end(); it != eIt; it++ )
+	for( Gadget *it : gadgets )
 	{
-		GadgetPtr gadget = *it;
+		Gadget* gadget = it;
 		while( gadget && gadget != this )
 		{
 			if( gadget == event.destinationGadget )
@@ -1484,16 +1568,16 @@ GadgetPtr ViewportGadget::updatedDragDestination( std::vector<GadgetPtr> &gadget
 	// keep the existing destination if it's also the source.
 	if( event.destinationGadget && event.destinationGadget == event.sourceGadget )
 	{
-		return event.destinationGadget;
+		return event.destinationGadget.get();
 	}
 
 	// and if that's not the case then give the drag source another chance
 	// to become the destination again.
 	if( event.sourceGadget )
 	{
-		if( dispatchEvent( event.sourceGadget, &Gadget::dragEnterSignal, event ) )
+		if( dispatchEvent( event.sourceGadget.get(), &Gadget::dragEnterSignal, event ) )
 		{
-			return event.sourceGadget;
+			return event.sourceGadget.get();
 		}
 	}
 
@@ -1507,7 +1591,7 @@ bool ViewportGadget::dragLeave( GadgetPtr gadget, const DragDropEvent &event )
 	{
 		GadgetPtr previousDestination = event.destinationGadget;
 		const_cast<DragDropEvent &>( event ).destinationGadget = nullptr;
-		dispatchEvent( previousDestination, &Gadget::dragLeaveSignal, event );
+		dispatchEvent( previousDestination.get(), &Gadget::dragLeaveSignal, event );
 	}
 	return true;
 }
@@ -1522,7 +1606,7 @@ bool ViewportGadget::drop( GadgetPtr gadget, const DragDropEvent &event )
 	{
 		if( event.destinationGadget )
 		{
-			return dispatchEvent( event.destinationGadget, &Gadget::dropSignal, event );
+			return dispatchEvent( event.destinationGadget.get(), &Gadget::dropSignal, event );
 		}
 		else
 		{
@@ -1551,7 +1635,7 @@ bool ViewportGadget::dragEnd( GadgetPtr gadget, const DragDropEvent &event )
 		m_dragTrackingIdleConnection.disconnect();
 		if( event.sourceGadget )
 		{
-			return dispatchEvent( event.sourceGadget, &Gadget::dragEndSignal, event );
+			return dispatchEvent( event.sourceGadget.get(), &Gadget::dragEndSignal, event );
 		}
 	}
 	return false;
@@ -1618,11 +1702,11 @@ void ViewportGadget::eventToGadgetSpace( ButtonEvent &event, Gadget *gadget )
 }
 
 template<typename Event, typename Signal>
-typename Signal::result_type ViewportGadget::dispatchEvent( std::vector<GadgetPtr> &gadgets, Signal &(Gadget::*signalGetter)(), const Event &event, GadgetPtr &handler )
+typename Signal::result_type ViewportGadget::dispatchEvent( std::vector<Gadget*> &gadgets, Signal &(Gadget::*signalGetter)(), const Event &event, Gadget* &handler )
 {
-	for( std::vector<GadgetPtr>::const_iterator it = gadgets.begin(), eIt = gadgets.end(); it != eIt; it++ )
+	for( Gadget* it : gadgets )
 	{
-		GadgetPtr gadget = *it;
+		Gadget* gadget = it;
 		if( !gadget->enabled() )
 		{
 			continue;
@@ -1642,12 +1726,12 @@ typename Signal::result_type ViewportGadget::dispatchEvent( std::vector<GadgetPt
 }
 
 template<typename Event, typename Signal>
-typename Signal::result_type ViewportGadget::dispatchEvent( GadgetPtr gadget, Signal &(Gadget::*signalGetter)(), const Event &event )
+typename Signal::result_type ViewportGadget::dispatchEvent( Gadget* gadget, Signal &(Gadget::*signalGetter)(), const Event &event )
 {
 	Event transformedEvent( event );
-	eventToGadgetSpace( transformedEvent, gadget.get() );
-	Signal &s = (gadget.get()->*signalGetter)();
-	return s( gadget.get(), transformedEvent );
+	eventToGadgetSpace( transformedEvent, gadget );
+	Signal &s = (gadget->*signalGetter)();
+	return s( gadget, transformedEvent );
 }
 
 //////////////////////////////////////////////////////////////////////////
@@ -1674,10 +1758,10 @@ ViewportGadget::SelectionScope::SelectionScope( const Imath::V3f &corner0InGadge
 	begin( viewportGadget, rasterRegion, gadget->fullTransform(), mode );
 }
 
-ViewportGadget::SelectionScope::SelectionScope( const ViewportGadget *viewportGadget, const Imath::V2f &rasterPosition, std::vector<IECoreGL::HitRecord> &selection, IECoreGL::Selector::Mode mode )
+ViewportGadget::SelectionScope::SelectionScope( const ViewportGadget *viewportGadget, const Imath::Box2f &rasterRegion, std::vector<IECoreGL::HitRecord> &selection, IECoreGL::Selector::Mode mode )
 	:	m_selection( selection )
 {
-	begin( viewportGadget, rasterPosition, M44f(), mode );
+	begin( viewportGadget, rasterRegion, M44f(), mode );
 }
 
 ViewportGadget::SelectionScope::~SelectionScope()