+++ Changes.md
@@ -5,11 +5,18 @@ API
 ---
 
 - GraphComponent : Inlined `children()` method, yielding 20-40% improvements in various child iteration benchmarks.
+- Gadget :
+  - Added `gadgetsAt()` overload which returns the gadgets rather than taking an output parameter by reference.
+  - Added `gadgetsAt()` overload taking a raster space region (rather than position) and an optional layer filter.
 
 Breaking Changes
 ----------------
 
 - FilteredChildIterator/FilteredRecursiveChildIterator : Removed all namespace-level typedefs, which were deprecated in Gaffer 0.59.0.0. Use the class-level typedefs instead, for example `Plug::Iterator` in place of `PlugIterator`.
+- Gadget :
+  - Moved `render()` and `renderRequestSignal()` to ViewportGadget.
+  - Made `select()` private.
+- ViewportGadget : Deprecated old `gadgetsAt()` signature. Please use the new form instead.
 
 0.60.0.0
 ========